refactor: publish domain event from event package
